refactor(web-chat): delete dead legacy code and fix migration test
